tapfreighter+tapsend: refactor output commitment creation
This commit changes the way output commitments are created.
Before this commit, the root split output (where we assumed the change
would go and/or where the passive assets were anchored to) was treated
differently than a simple split root output:
We always started with the input commitment, cloned it, removed the
active asset from the commitment, then added the output asset. And
finally we also committed the passive assets to that output.
But because the input commitment minus the active asset is equal to the
passive assets, we can simplify this whole procedure if we just assume
that the passive assets are in the list of the given virtual packets.